Limit term updates to 10/s.
The current code pushes an update to the term for everything that comes
in; end result, this can result in a lot of time being spent on
just updating the term, time that should be allocated elsewhere.

For a local hot-cache test, this reduces a full generation scan from 8.8s
to 7.05s, so it's not neglible.